Refactor code for improved organization: Adjust namespaces for Feature classes and relocate utility classes to Utils.
