+++ stardew-access/API.cs
@@ -1,4 +1,5 @@
 using Microsoft.Xna.Framework;
+using stardew_access.Features;
 using stardew_access.Utils;
 
 namespace stardew_access
