+++ stardew-access/Features/Radar.cs
@@ -1,8 +1,9 @@
 using Microsoft.Xna.Framework;
+using stardew_access.Utils;
 using StardewValley;
 using StardewValley.Objects;
 
-namespace stardew_access.Utils
+namespace stardew_access.Features
 {
     public class Radar
     {
