+++ stardew-access/Utils/GameStateNarrator.cs
