+++ stardew-access/Utils/StaticTiles.cs
